Change projectile directory name
Change the projectile directory name to have a capital P. It was originally with a capital P and I tried to change it to be consistent with the other directories in our project. However, Git did not like this change for some reason and doesn't change the directory name when I switch branches. It only changes the contents inside each file (so the imports) to have the old name, which throws a bunch of errors.