Implement ready to fire LEDs
Make the changes necessary to implement ready to fire LEDs, including logic upgrades for the shooter subsystem to allow RPMs to be set without them being considered "ready to fire". This is useful in the case that we set the shooter to a pre-determined RPM because we aren't able to see a target.